Move type_info away from node_modules
