+++ radix-engine/src/blueprints/resource/auth_zone/blueprint.rs
@@ -2,9 +2,9 @@ use crate::blueprints::resource::ComposedProof;
 use crate::errors::*;
 use crate::kernel::kernel_api::{KernelNodeApi, KernelSubstateApi};
 use crate::system::node_init::type_info_partition;
-use crate::system::node_modules::type_info::TypeInfoSubstate;
 use crate::system::system_callback::SystemLockData;
 use crate::system::system_modules::auth::{Authorization, AuthorizationCheckResult};
+use crate::system::type_info::TypeInfoSubstate;
 use crate::types::*;
 use native_sdk::resource::NativeProof;
 use radix_engine_interface::api::{ClientApi, LockFlags, ObjectModuleId, OBJECT_HANDLE_SELF};
