+++ radix-engine/src/blueprints/transaction_processor/tx_processor.rs
@@ -4,9 +4,9 @@ use crate::errors::RuntimeError;
 use crate::kernel::kernel_api::KernelNodeApi;
 use crate::kernel::kernel_api::KernelSubstateApi;
 use crate::system::node_init::type_info_partition;
-use crate::system::node_modules::type_info::TypeInfoBlueprint;
-use crate::system::node_modules::type_info::TypeInfoSubstate;
 use crate::system::system::FieldSubstate;
+use crate::system::type_info::TypeInfoBlueprint;
+use crate::system::type_info::TypeInfoSubstate;
 use crate::types::*;
 use native_sdk::resource::NativeFungibleBucket;
 use native_sdk::resource::NativeNonFungibleBucket;
