+++ radix-engine/src/system/node_init.rs
@@ -1,4 +1,4 @@
-use crate::system::node_modules::type_info::TypeInfoSubstate;
+use crate::system::type_info::TypeInfoSubstate;
 use crate::types::*;
 use radix_engine_interface::types::SubstateKey;
