+++ radix-engine/src/transaction/state_update_summary.rs
@@ -9,9 +9,9 @@ use radix_engine_store_interface::{
 };
 use sbor::rust::prelude::*;
 
-use crate::system::node_modules::type_info::TypeInfoSubstate;
 use crate::system::system::FieldSubstate;
 use crate::system::system_db_reader::SystemDatabaseReader;
+use crate::system::type_info::TypeInfoSubstate;
 use crate::track::TrackedSubstateValue;
 use crate::track::{TrackedNode, Write};
